+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - storage-pool list reporting offline nodes even tough everything is connected
 - SED setup using same password for all drives
+- Fixed stuck deletion of resource during failed / failing resize
 
 ## [1.29.0] - 2024-07-31